add _disable_strategy, test=develop
